+++ packages/components/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 9.1.7
+
+### Patch Changes
+
+- Fix build
+- Updated dependencies [undefined]
+  - @heathmont/moon-assets@9.1.7
+  - @heathmont/moon-icons@9.1.7
+  - @heathmont/moon-themes@9.1.7
+  - @heathmont/moon-utils@9.1.7
+
 ## 9.1.6
 
 ### Patch Changes