+++ workspaces/themes/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @heathmont/moon-themes-tw
 
+## 9.1.7
+
+### Patch Changes
+
+- Fix build
+
 ## 9.1.6
 
 ### Patch Changes